requirements.txt
langchain
streamlit
langchain-community
pinecone-client
openai
tiktoken
nest_asyncio